[Diplom_2]added changes in data.py and remove main.py
