+++ data.py
@@ -1,17 +1,6 @@
-from faker import Faker
 
 
 class User:
-
-    @staticmethod
-    def create_user():
-        fake = Faker()
-
-        reg_data = {
-            "email": fake.email(),
-            "password": fake.password(),
-            "name": fake.name()}
-        return reg_data
 
     data_correct = {
         "email": 'kotishka2012@yandex.ru',
