+++ helpers.py
@@ -1,0 +1,11 @@
+from faker import Faker
+
+
+def create_user():
+    fake = Faker()
+
+    reg_data = {
+        "email": fake.email(),
+        "password": fake.password(),
+        "name": fake.name()}
+    return reg_data
